Create api model to fetch data.
